refactor(JwtCookieManager): 쿠키 생성의 책임을 컨트롤러에서 분리
